fix: Continue simulation until all philos full
This is to fix that someone dies with very tight timings at the end with uneven amount of philos, simply because of the perfect spacing that enables usually impossible timings.

Some philos will print one more eat message than num_each_philo_must_eat. But that should be fine with the subject.
Also, they don't finish their meal, and the meal counter only goes up once the meal has been fully finished, so that's another argument for it being fine.